06-stock/requirements.txt
newsapi-python
requests
python-dotenv
black
mypy
twilio